Add Michael Jordan to NBA Jam! Rewind Edition!
nbajamtr

Personal thanks to DemiGodX for working on this amazing hack, as well as Mahoneyt944, Zoltan45...And, of course, those who worked on the original game!

Support with this hacked revision also incorporates and is supplemented by Shaquille O'Neal, Mortal Kombat Characters (complete with fatalities!), as well as playable cheerleaders, smashable backboards, and other such fun stuff that the NBA would have never wanted in this game:)

Enjoy!